pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "A RAG-based AI bot for database procedure application flows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.6.3",
    "fastapi>=0.115.8",
    "google-generativeai>=0.8.4",
    "langchain>=0.3.17",
    "langchain-community>=0.3.17",
    "langchain-google-genai>=2.0.9",
    "langchain-text-splitters>=0.3.6",
    "pandas>=2.3.3",
    "pydantic>=2.10.7",
    "pydantic-settings>=2.7.1",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.38",
    "streamlit>=1.55.0",
    "uvicorn>=0.34.0",
]

[tool.uv]
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'x86_64'",
    "sys_platform == 'darwin' and platform_machine == 'arm64'",
    "sys_platform == 'linux'",
]
